v2/requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-cors==4.0.0
flask-limiter==3.7.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flasgger==0.9.7.1
marshmallow==3.21.3
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.3
sqlalchemy==2.0.30
rdkit==2024.3.5
torch==2.3.1
transformers==4.42.0
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
gunicorn==22.0.0
python-dotenv==1.0.1
click==8.1.7
